+++ src/main/java/ar/com/chepps/Companny/entity/ProductoDetalle.java
@@ -2,14 +2,16 @@ package ar.com.chepps.Companny.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table (name = "ProductoDetalle")
 public class ProductoDetalle {
     @Id
     @GeneratedValue (strategy = GenerationType.IDENTITY)
     private Long idProductoDetalle;
-    private double stockActual;
-    private double stockMinimo;
+    private BigDecimal stockActual;
+    private BigDecimal stockMinimo;
 
     public ProductoDetalle() {
     }
@@ -18,19 +20,19 @@ public class ProductoDetalle {
         return idProductoDetalle;
     }
 
-    public double getStockActual() {
+    public BigDecimal getStockActual() {
         return stockActual;
     }
 
-    public void setStockActual(double stockActual) {
+    public void setStockActual(BigDecimal stockActual) {
         this.stockActual = stockActual;
     }
 
-    public double getStockMinimo() {
+    public BigDecimal getStockMinimo() {
         return stockMinimo;
     }
 
-    public void setStockMinimo(double stockMinimo) {
+    public void setStockMinimo(BigDecimal stockMinimo) {
         this.stockMinimo = stockMinimo;
     }
 }
